fixes issues
