+++ mlbstatsapi/models/drafts/__init__.py
@@ -1,2 +1,1 @@
 ﻿from .rounds import Round
-from .attributes import Home, College
